Split util.rkt into basic.rkt and adv.rkt
This solves dependency issues that were occuring:

basic.rkt now serves the original purpose of util.rkt. It is included by
data.rkt, and provides basic, general functions which have nothing in
particular to do with this project.

adv.rkt includes data.rkt, constants.rkt, and cell.rkt; it provides functions
which are specific to this project, and potentially rely on the data
structures.